algo/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
python-multipart>=0.0.5
sse-starlette>=1.6.0
langchain>=0.0.350
langchain-community>=0.0.10
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
soundfile>=0.12.0
librosa>=0.10.0
loguru>=0.7.0

# 语音处理依赖
openai>=1.0.0
azure-cognitiveservices-speech>=1.34.0
edge-tts>=6.1.0
SpeechRecognition>=3.10.0
pydub>=0.25.0
webrtcvad>=2.0.10
